Fix the GetIdentityContent to use the start and end heights and also move it into its own directory.
